fix: resolve circular import + unhandled refresh + partial test mock in PAT panel
- Move CopyBlock out of SettingsPage.tsx into components/CopyBlock.tsx so
  components/ never imports from pages/ (page -> component -> page cycle).
- TokensPanel.refresh() now catches listTokens() failures and surfaces an
  error instead of leaving an unhandled rejection with a stuck list.
- SettingsPage.presence.test.tsx's @/api/tokens mock now stubs all three
  exports, not just listTokens.
